requirements.txt
fastapi==0.135.3
SQLAlchemy==2.0.49
alembic==1.18.4
psycopg2-binary==2.9.11
python-dotenv==1.2.2
uvicorn==0.44.0
